% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FillNulls.R
\name{FillNulls}
\alias{FillNulls}
\title{Fills Null values with NA}
\usage{
FillNulls(y)
}
\arguments{
\item{y}{json object with null values}
}
\value{
json object
}
\description{
Fill the null values in the training and test datasets with NA
}

