% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCutoff.R
\name{FindCutoff}
\alias{FindCutoff}
\title{Find the cuttoff for binary variables}
\usage{
FindCutoff(objective, lever, range)
}
\arguments{
\item{objective}{vector of classes, should be binary 0/1 or T/F}

\item{lever}{numeric vector to}

\item{range}{range of values used as cutoffs}
}
\value{
vector of percentages by class
}
\description{
Find the cutoff point when converting a numeric vector to binary. The
function calculates the share of a instances in a class given a
cutoff point.
}
