% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomActsofPizza.R
\docType{data}
\name{justInCase}
\alias{justInCase}
\title{Extraneous data points}
\description{
This dataset is the product of Exploratory Analysis where
the raw training dataset was downloaded and processed. Call
\code{vignette("ExploratoryAnalysis")} to see the documentation.
}

