% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeModels.R
\name{MergeModels}
\alias{MergeModels}
\title{To merge the models into singe predictions}
\usage{
Preds< cbind(1:5,6:10)
MergeModels(Preds, c(.4,.6)) # different weights
MergeModels(Preds) # Equal weights
}
\arguments{
\item{x}{matrix or dataframe of predictions from singe models}

\item{weights}{vector of weights to apply to the dataframe}
}
\value{
single prediction for each observation
}
\description{
Takes a weighted average of the prediction scores to create and ensembled model.
}

