% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{PrepareData}
\alias{PrepareData}
\title{Convert to tabular format}
\usage{
PrepareData(y)
}
\arguments{
\item{y}{either training or test datasets in json format}
}
\value{
dataframe
}
\description{
Prepare a tabular dataset from json
}

