% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessText.R
\name{ProcessText}
\alias{ProcessText}
\title{Preprocess text fields}
\usage{
ProcessText(x, remove = NULL)
}
\arguments{
\item{x}{character for preprocessing}

\item{remove}{a vector of additional words to remove}
}
\value{
Document term matrix
}
\description{
To preprocess text fields so they can be used for further analysis. The
preprocessing follows these steps:
\itemize{
\item Convert to corpus
\item Remove capitalized letters
\item Convert to plain text
\item Remove punctuation
\item Remove stopwords
\item Stem words
\item Convert to document term matrix
}
}

