% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleprint.R
\name{simpleprint}
\alias{simpleprint}
\title{Simple print function}
\usage{
simpleprint(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Prints the contents of a dataframe without column or row names
}

