% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomActsofPizza.R
\docType{data}
\name{test}
\alias{test}
\title{Test dataset}
\format{dataframe with 1631 rows and 24 variables.
\itemize{
\item giver_username_if_known
\item request_id
\item requester_account_age_in_days_at_request
\item requester_days_since_first_post_on_raop_at_request
\item requester_number_of_comments_at_request
\item requester_number_of_comments_in_raop_at_request
\item requester_number_of_posts_at_request
\item requester_number_of_posts_on_raop_at_request
\item requester_number_of_subreddits_at_request
\item requester_subreddits_at_request
\item requester_upvotes_plus_downvotes_at_request
\item requester_username
\item Year
\item weekday
\item Image
\item Acct.Age
\item BnRAOP
\item Words
\item Word.bin
}}
\description{
This dataset is the product of Exploratory Analysis where
he raw training dataset was downloaded and processed. Call
\code{vignette("ExploratoryAnalysis")} to see the documentation.
}

