% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomActsofPizza.R
\docType{data}
\name{train}
\alias{train}
\title{Training dataset}
\format{dataframe with 4040 rows.  There are 17 primary
variables plus additional variables to cound the number of
times a word occurs in a request.
\itemize{
\item giver_username_if_known
\item account_age_in_days
\item days_since_first_post_on_raop
\item number_of_comments
\item number_of_comments_in_raop
\item number_of_posts
\item number_of_posts_on_raop
\item number_of_subreddits
\item upvotes_plus_downvotes
\item received_pizza
\item Year
\item weekday
\item Image
\item Acct.Age
\item BnRAOP
\item Words
\item Word.bin
}}
\description{
This dataset is the product of Exploratory Analysis where
he raw training dataset was downloaded and processed. Call
\code{vignette("ExploratoryAnalysis")} to see the documentation.
}
